+++ Source/AmrMesh/CD_AmrMesh.cpp
@@ -1149,7 +1149,8 @@ AmrMesh::buildGrids(const Vector<IntVectSet>& a_tags, const int a_lmin, const in
   // baseLevel is the coarsest level which does not change. topLevel is the finest level where we have tags. We should never
   // have tags on max_amr_depth, and we make that restriction here.
   const int baseLevel = std::max(0, a_lmin - 1);
-  const int topLevel  = (m_finestLevel == m_maxAmrDepth) ? m_finestLevel - 1 : a_tags.size() - 1;
+  const int topLevel  = (m_finestLevel == m_maxAmrDepth) ? static_cast<int>(m_finestLevel) - 1
+                                                         : static_cast<int>(a_tags.size()) - 1;
 
   // New and old grid boxes
   Vector<Vector<Box>> newBoxes(1 + topLevel);
@@ -1244,7 +1245,7 @@ AmrMesh::buildGrids(const Vector<IntVectSet>& a_tags, const int a_lmin, const in
     Vector<long int>   boxLoads(levelBoxes.size());
 
     for (int ibox = 0; ibox < levelBoxes.size(); ibox++) {
-      boxLoads[ibox] = levelBoxes[ibox].numPts();
+      boxLoads[ibox] = static_cast<long>(levelBoxes[ibox].numPts());
     }
 
     // Load balance this grid -- assign grid subsets to the least loaded rank.
@@ -2531,7 +2532,7 @@ AmrMesh::parseRefinementRatios()
   ParmParse   pp("AmrMesh");
   Vector<int> ratios;
   ratios.resize(pp.countval("ref_rat"));
-  pp.getarr("ref_rat", ratios, 0, ratios.size());
+  pp.getarr("ref_rat", ratios, 0, static_cast<int>(ratios.size()));
 
   // Pad with whatever was last specified if user didn't supply enough refinement factors
   while (ratios.size() < m_maxAmrDepth) {
