+++ Source/AmrMesh/CD_EBCoarAve.cpp
@@ -157,7 +157,7 @@ EBCoarAve::defineCellStencils() noexcept
     auto buildStencils = [&](const VolIndex& coarVoF) -> void {
       const Real             kappaC      = ebisBoxCoar.volFrac(coarVoF);
       const Vector<VolIndex> fineVoFs    = ebislCoar.refine(coarVoF, m_refRat, din);
-      const int              numFineVoFs = fineVoFs.size();
+      const int              numFineVoFs = static_cast<int>(fineVoFs.size());
 
       VoFStencil& arithSten = arithmeticStencils(coarVoF, 0);
       VoFStencil& harmSten  = harmonicStencils(coarVoF, 0);
@@ -239,7 +239,7 @@ EBCoarAve::defineFaceStencils() noexcept
       auto buildStencils = [&](const FaceIndex& coarFace) -> void {
         const Real              areaCoar     = ebisBoxCoar.areaFrac(coarFace);
         const Vector<FaceIndex> fineFaces    = ebislCoar.refine(coarFace, m_refRat, din);
-        const int               numFineFaces = fineFaces.size();
+        const int               numFineFaces = static_cast<int>(fineFaces.size());
 
         FaceStencil& arithSten = arithmeticStencils(coarFace, 0);
         FaceStencil& harmSten  = harmonicStencils(coarFace, 0);
@@ -329,7 +329,7 @@ EBCoarAve::defineEBStencils() noexcept
           fineVoFs.push_back(refinedVoFs[i]);
         }
       }
-      const int numFineVoFs = fineVoFs.size();
+      const int numFineVoFs = static_cast<int>(fineVoFs.size());
 
       if (numFineVoFs > 0) {
         for (int ifine = 0; ifine < numFineVoFs; ifine++) {
@@ -511,7 +511,7 @@ EBCoarAve::harmonicAverage(EBCellFAB&       a_coarData,
 
   // Regular cells
   const Box  refiBox    = Box(IntVect::Zero, (m_refRat - 1) * IntVect::Unit);
-  const Real numPerCoar = refiBox.numPts();
+  const Real numPerCoar = static_cast<double>(refiBox.numPts());
 
   FArrayBox&       coarDataReg = a_coarData.getFArrayBox();
   const FArrayBox& fineDataReg = a_fineData.getFArrayBox();
